requirements.txt
django>=3.2,<3.3
django-pwa>=1.0.10,<1.1
django-bootstrap4>=3.0.1,<3.1
pillow>=8.3.2,<8.4
django-bootstrap-datepicker-plus>=3.0.5,<3.1
django-crispy-forms>=1.13.0,<1.14
django-mathfilters >=1.0.0,<1.1
requests>=2.26.0,<2.27
django-extra-views>=0.14.0,<0.15
gunicorn>=20.1.0,<20.2